RoomDetailView에서 방 나가기, 동승자, 채팅 버튼 컴포넌트 추가 및 스타일 수정; PartyEmpty 컴포넌트의 텍스트 색상 변경; 불필요한 코드 제거
