+++ components/my_party/icon_styles.tsx
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  button: {
+    alignContent: "center",
+  },
+  buttonText: {
+    fontSize: 14,
+  },
+});
